+++ tiger-rbel/src/test/java/de/gematik/rbellogger/data/RbelJexlTest.java
@@ -4,22 +4,19 @@
 
 package de.gematik.rbellogger.data;
 
-import de.gematik.rbellogger.RbelLogger;
-import de.gematik.rbellogger.RbelOptions;
-import de.gematik.rbellogger.converter.RbelJexlExecutor;
-import java.time.ZonedDateTime;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-
-import java.io.IOException;
-import java.util.Optional;
-
 import static de.gematik.rbellogger.TestUtils.localhostWithPort;
 import static de.gematik.rbellogger.TestUtils.readCurlFromFileWithCorrectedLineBreaks;
 import static org.assertj.core.api.Assertions.assertThat;
+import de.gematik.rbellogger.RbelLogger;
+import de.gematik.rbellogger.RbelOptions;
+import de.gematik.rbellogger.converter.RbelJexlExecutor;
+import java.io.IOException;
+import java.time.ZonedDateTime;
+import java.util.Optional;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
-public class RbelJexlTest {
+class RbelJexlTest {
 
     private RbelElement response;
     private RbelElement request;
@@ -40,7 +37,7 @@ public class RbelJexlTest {
     }
 
     @Test
-    public void checkRequestMapElements() {
+    void checkRequestMapElements() {
         assertThat(jexlExecutor.matchesAsJexlExpression(
             request, "isRequest", Optional.empty()))
             .isTrue();
@@ -53,7 +50,7 @@ public class RbelJexlTest {
     }
 
     @Test
-    public void checkResponseStatusCode() {
+    void checkResponseStatusCode() {
         assertThat(jexlExecutor.matchesAsJexlExpression(
             request, "response.statusCode == 200", Optional.empty()))
             .isTrue();
@@ -72,7 +69,7 @@ public class RbelJexlTest {
     }
 
     @Test
-    public void checkResponseMapElements() {
+    void checkResponseMapElements() {
         assertThat(jexlExecutor.matchesAsJexlExpression(
             response, "isRequest == false", Optional.empty()))
             .isTrue();
@@ -85,7 +82,7 @@ public class RbelJexlTest {
     }
 
     @Test
-    public void checkJexlParsingForDoubleHeaders() throws IOException {
+    void checkJexlParsingForDoubleHeaders() throws IOException {
         RbelElement doubleHeaderMessage = RbelLogger.build().getRbelConverter()
             .parseMessage(readCurlFromFileWithCorrectedLineBreaks
                 ("src/test/resources/sampleMessages/doubleHeader.curl").getBytes(), null, null, Optional.of(ZonedDateTime.now()));
@@ -96,7 +93,7 @@ public class RbelJexlTest {
     }
 
     @Test
-    public void shouldFindReceiverPort() throws IOException {
+    void shouldFindReceiverPort() throws IOException {
         RbelElement request = RbelLogger.build().getRbelConverter().parseMessage(
             readCurlFromFileWithCorrectedLineBreaks("src/test/resources/sampleMessages/getRequest.curl").getBytes(),
             localhostWithPort(44444), localhostWithPort(5432), Optional.of(ZonedDateTime.now()));
@@ -107,7 +104,7 @@ public class RbelJexlTest {
     }
 
     @Test
-    public void checkMatchTextExpression() {
+    void checkMatchTextExpression() {
         assertThat(jexlExecutor.matchAsTextExpression(
             request, "localhost"))
             .isTrue();
@@ -120,7 +117,7 @@ public class RbelJexlTest {
     }
 
     @Test
-    public void checkMatchRegexExpression() {
+    void checkMatchRegexExpression() {
         assertThat(jexlExecutor.matchAsTextExpression(
             request, "\\w+host"))
             .isTrue();
@@ -136,5 +133,11 @@ public class RbelJexlTest {
         assertThat(jexlExecutor.matchAsTextExpression(
             request, "Keep-[a-zA-Z0-9_]+Alive"))
             .isFalse();
+    }
+
+    @Test
+    void testValueFacetComparison() {
+        assertThat(jexlExecutor.matchesAsJexlExpression(response, "$..signature.isValid == 'true'"))
+            .isTrue();
     }
 }